backend/requirements.txt
#temporary 

fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-multipart==0.0.6
email-validator==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
authlib==1.2.1
redis==5.0.1
fastapi-cache2[redis]==0.2.1
fastapi-limiter==0.1.5
